+++ src/components/header/Header.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { StHeader } from './style';
+
+// 홈 화면 헤더 영역
+const Header = () => {
+  return (
+    <StHeader>
+      <h3>TODOLIST</h3>
+      <h3>React-redux</h3>
+    </StHeader>
+  );
+};
+
+export default Header;
