+++ src/components/header/style.js
@@ -1,17 +1,4 @@
-import React from 'react';
 import styled from 'styled-components';
-
-const Header = () => {
-  return (
-    // 홈 화면 헤더 영역
-    <StHeader>
-      <h3>TODOLIST</h3>
-      <h3>React-redux</h3>
-    </StHeader>
-  );
-};
-
-export default Header;
 
 const StHeader = styled.div`
   height: 50px;
@@ -24,3 +11,5 @@ const StHeader = styled.div`
   background-color: #000000;
   color: white;
 `;
+
+export { StHeader };
