+++ src/components/layout/style.js
@@ -1,12 +1,4 @@
-// src/shared/Layout.js
 import styled from 'styled-components';
-import React from 'react';
-
-function Layout({ children }) {
-  return <StLayout>{children}</StLayout>;
-}
-
-export default Layout;
 
 const StLayout = styled.div`
   max-width: 1200px;
@@ -16,3 +8,5 @@ const StLayout = styled.div`
   background-color: white;
   box-shadow: -5px 20px 30px rgba(0, 0, 0, 0.2);
 `;
+
+export { StLayout };
