+++ src/components/todoForm/TodoForm.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { addTodo } from '../../redux/modules/todos';
 import { v4 as uuidv4 } from 'uuid';
-import styled from 'styled-components';
+import { StForm, StInputContainer, StLabeledInput, StFormBtn } from './styles';
 
 const TodoForm = () => {
   // 제목, 내용 컴포넌트 state 설정
@@ -77,38 +77,3 @@ const TodoForm = () => {
 };
 
 export default TodoForm;
-
-const StForm = styled.form`
-  padding-top: 20px;
-  background-color: #eee;
-  display: flex;
-  justify-content: center;
-`;
-
-const StInputContainer = styled.div`
-  width: 80%;
-  margin-bottom: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const StLabeledInput = styled.input`
-  width: 700px;
-  height: 30px;
-  padding: 15px;
-  border: none;
-  background: #ffffff;
-  border-radius: 100px;
-  margin-bottom: 10px;
-  margin-right: 20px;
-`;
-
-const StFormBtn = styled.button`
-  width: 65px;
-  height: 30px;
-  font-weight: 500;
-  border-radius: 100px;
-  background-color: #005cb2;
-  color: #ffffff;
-`;
